backend/services/data_ingestion/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2
paho-mqtt
requests
